fix mov.l1.llb, mov.llb.l1, icmov,m, pld copy len variable data type

+++ riscv/insns/icmov_m.h
@@ -12,10 +12,10 @@ check_traps_icmov_m(e_size);
 auto dst_coreId = DST_CORE_ID;
 
 //src shape
-unsigned short col = MTE_SHAPE_COLUMN;
-unsigned short row = MTE_SHAPE_ROW; 
-unsigned short copy_stride_rs1 = (MTE_STRIDE_RS1 ? MTE_STRIDE_RS1 : col) * e_size;
-unsigned short copy_stride_rd = (MTE_STRIDE_RD ? MTE_STRIDE_RD : col) * e_size;
+uint16_t col = MTE_SHAPE_COLUMN;
+uint16_t row = MTE_SHAPE_ROW; 
+uint32_t copy_stride_rs1 = (MTE_STRIDE_RS1 ? MTE_STRIDE_RS1 : col) * e_size;
+uint32_t copy_stride_rd = (MTE_STRIDE_RD ? MTE_STRIDE_RD : col) * e_size;
 
 auto src = (uint8_t*)MMU.get_phy_addr(RS1);
 auto dst = (uint8_t*)p->get_sim()->local_addr_to_mem_by_id_cluster(zext_xlen(RD), dst_coreId);
