+++ riscv/insns/pld.h
@@ -15,7 +15,7 @@ reg_t rs2 = RS2;
 p->run_async([p, insn, pc, xlen, addr, dst_addr, e_size, rs2]() {
   uint8_t* src = (uint8_t*)p->get_sim()->addr_to_mem(addr);
   uint8_t* dst = (uint8_t*)MMU.get_phy_addr(dst_addr);
-  unsigned int core_map = (unsigned int)rs2;
+  uint32_t core_map = (uint32_t)rs2;
 
   // do sync for pld
   p->pld(core_map);
@@ -26,11 +26,11 @@ p->run_async([p, insn, pc, xlen, addr, dst_addr, e_size, rs2]() {
   #endif
 
   //src shape
-  unsigned short col = MTE_SHAPE_COLUMN;
-  unsigned short row = MTE_SHAPE_ROW; 
-  unsigned short copy_stride_rs1 = 0;
-  unsigned short copy_stride_rd = 0;
-  unsigned int core_id = p->get_csr(CSR_TID);
+  uint16_t col = MTE_SHAPE_COLUMN;
+  uint16_t row = MTE_SHAPE_ROW; 
+  uint32_t copy_stride_rs1 = 0;
+  uint32_t copy_stride_rd = 0;
+  uint32_t core_id = p->get_csr(CSR_TID);
 
   if (core_map != 0 && (core_map & (1 << core_id))) {
     copy_stride_rs1 = (MTE_STRIDE_RS1 ? MTE_STRIDE_RS1 : col) * e_size;
